Consistently track what is being selected and deselected
Originally this would only tell the outside what was clicked last. WWT was just layering items. Effectively, multiple items where "selected". The behavior I believe was desired is the deselect the previous item. This is done by using "single-select=true" and @deselect="onItemDeselected" which lowers the opacity to 0. The original behavior can be mimmiced using "single-select=false" :highlight-last-only=true and @delselect unassigned (though now it explicitly shows it as selected in the gallery).

The other solution would have been to modify onItemSelected to hide everything else, but I think it better to